Python 3 bump

+++ src/openbandparams/examples/Plot_Band_Offset_vs_Composition_of_Ternary.py
@@ -56,8 +56,8 @@ for x, y, label in zip(x, y, label):
 # plot the ternaries
 indices = numpy.arange(100)
 fractions = numpy.linspace(0, 1, 100)
-x = numpy.empty(100, dtype=numpy.float)
-y = numpy.empty(100, dtype=numpy.float)
+x = numpy.empty(100, dtype=numpy.float32)
+y = numpy.empty(100, dtype=numpy.float32)
 for tern in [alloy]:
     for i, f in zip(indices, fractions):
         instance = tern(x=f)
@@ -84,8 +84,8 @@ for x, y, label in zip(x, y, label):
 # plot the ternaries
 indices = numpy.arange(100)
 fractions = numpy.linspace(0, 1, 100)
-x = numpy.empty(100, dtype=numpy.float)
-y = numpy.empty(100, dtype=numpy.float)
+x = numpy.empty(100, dtype=numpy.float32)
+y = numpy.empty(100, dtype=numpy.float32)
 for tern in [alloy]:
     for i, f in zip(indices, fractions):
         instance = tern(x=f)
